config moved into the right places
