+++ conf_templates/admin_settings.json
