+++ conf_templates/challenges.json
